[NCL-7921] Add annotations to influence request
This commit adds the 'annotations' field to the Request object. This
isn't part of the actual request, but can be used to influence how the
Request is built in the client processing it.

For example, the annotations could indicate that we need to add the
authentication header in the request when it is processed by the client
by defining the key 'AUTHENTICATION_HEADER_NEEDED' and leaving the value
as an empty list.

It is a map with a key of String and as value a list of Strings. We
didn't use a MultiValuedMap since Jackson cannot deserialize it properly
out of the box.